Spellcheck.py is now redundant
For now, spellcheck is redundant, grammar_check will be taking over its functionality as it checks for both spelling mistakes and grammar issues. For now, this does make exceptions.tsv unusable, but that can be worked on later. I updated grammar_check to better format how the mistakes/corrections are returned as well as made it return the corrected text. Keywords is also updated to correctly handle dictionary additions and subtractions, as well handle the now corrected text, although it is still a bit incorrect in cases of words that contain separate distinct words in them.