Rough Grade added
a rough grade file was added, all of the grammar checking, Keyword checks, model comparisons, and otherwise should happen through it instead of main. As of now, there are two weights, grammar weight and key word weight, these are a measure of how much they effect the overall grade. feel free to play with them.