Model Inputs Updated
I changed the model inputs from being all essays to just one essay set, to make grading against the essay set more accurate. I also made it so that the correct score is calculated across any essay set, since the domain1_score isn't simply a sum of raters 1-3 across all essays sets, so the sum is manually calculated now. I also added in the ability to get the max and min possible scores for the currently worked on essay set, so that's there, even if it is unused.