Altered Model input
Before, the model would take in all essays to train and test against, which is wrong since the essay_sets are different, which leads to a situation where an essay about a book being graded against a set of essays about computers, which is wrong. Essays of a set should only be compared to essays of the same set. I also changed it to better calculate the total rater score of the training set essays, since before domain1_score was used, which while represents the sum of rater 1 and 2 in essay set 1, is instead only rater 1's score in essay set 2, but now this is more accurate. The ability to get the max and min grade provided by the model is also added, albeit unused currently. More work needs to be done to make the model return a more accurate score.