More Accurate Models (Hopefully)
The Score model has a 0.85 kappa score and the feedback models have a kappa score of about 0.9. It does take a while to build the se models, but this shouldn't be a problem as they only need to be built once. I also added the ability to specify the n folds and epochs per model, so there is that now. I do say that these models are hopefully accurate as there is a chance that these high scores are due to over fitting, but it is hard to check for that issue