Added NumPy Docstring
I am beginning to add NumPy style Docstring to every method and class in every file, starting with grade.py, which isn't completely done, but I do have the worst of it done. I also changed grade initialization to have only style instead of style and style_path as parameters. you can give a filepath or a dictionary to the style parameter now. Same goes for get_grade, instead of either text or filepath as an input, now there is only text, which can take both text or a filepath.
A bug was fixed with the format skeleton key.
All grade.py variables are now explicitly private, use the get methods provided