More Docstrings, but also slots
If you don't know what a slot is, its a way of telling python all of your planned class variables so it can better optimize the program when running. Nothing really big.
On another note, format.py is fully documented now, and has been fixed. The issue was that certain fields didn't exist when format thought that they would, but this issue has been caught now, at least in the test.docx I was given.