Added my code to updated version of website
In app.py, I changed how to processSave function works so that we can store all the necessary data for the essays.
I also moved the email field to be above the submit button, because that makes more sense from a UX perspective.
I also combined Evaluate, Save, and Email into one button, because each essay should be saved to the database anyways, so it doesn't make sense to make the user click three buttons. Finally I added a results route to display the results when a user clicks on their link

email_user.py now sends a link instead of just basic information. The password for the email address was removed from this file, if you need it let me know.

I also created iga.sh. The .bat file only works on windows, and our AWS server uses Linux, so it needs a .sh file instead.