Better File Paths
OK, so now if you want to initialize Grade(), give it a 'start' parameter that is a file path to src. If you want to change the destination of a specific file path, get the file path dictionary now provided in grade.py and after changing the relevant path, pass the dictionary to Grade(). I also changed Grade to have more specific exceptions, being KeyError and FileNotFoundError, and grade.py will no longer catch them, as it is technically the responsibility of the one who passed the poor parameters to handle the exceptions, not grade.py's. I have updated the grade.py doc strings to match the changes. The only change I made to app.py was updating the Grade() initialization.