NOTICE: Changing File Directory
Instead of pointing to the src folder, grade should now be set to point to the data folder. model_weights will be stored in the data folder as a consequence of this.
Otherwise, there were apparently some serious issues with keywords that went unnoticed until now, so now not only is keywords more correct, it now raises the correct specific exceptions, detailed in the docstrings. test_cases.py now fully covers both grammar_check and keywords, at least for now.